Btc dw-bridge renamed packaging
