+++ deploy/docker-compose-full.yml
@@ -48,7 +48,7 @@ services:
     - IROHA_HOST=d3-iroha
     - IROHA_PORT=50051
     - JAR_NAME=btc-dw-bridge.jar
-    - CLASS=dwbridge.btc.BtcDepositWithdrawalMain
+    - CLASS=com.d3.btc.dwbridge.BtcDepositWithdrawalMain
     networks:
     - d3-network
